Refactor format functions to new utils file
