+++ containers/ecr-viewer/src/app/tests/format-service.test.tsx
@@ -1,0 +1,32 @@
+import { formatDate } from "@/app/format-service";
+
+describe("Format Date", () => {
+  it("should return the correct formatted date", () => {
+    const inputDate = "2023-01-15";
+    const expectedDate = "01/15/2023";
+
+    const result = formatDate(inputDate);
+    expect(result).toEqual(expectedDate);
+  });
+
+  it("should return N/A if provided date is an empty string", () => {
+    const inputDate = "";
+
+    const result = formatDate(inputDate);
+    expect(result).toBeUndefined();
+  });
+
+  it("should return N/A if provided date is undefined", () => {
+    const inputDate = undefined;
+
+    const result = formatDate(inputDate as any);
+    expect(result).toBeUndefined();
+  });
+
+  it("should return N/A if provided date is null", () => {
+    const inputDate = null;
+
+    const result = formatDate(inputDate as any);
+    expect(result).toBeUndefined();
+  });
+});
