+++ containers/ecr-viewer/src/app/utils.tsx
@@ -1,3 +1,5 @@
+import React from "react";
+import * as dateFns from "date-fns";
 import {
   Bundle,
   Condition,
@@ -7,15 +9,24 @@ import {
   Procedure,
 } from "fhir/r4";
 import { evaluate } from "fhirpath";
-import { Table } from "@trussworks/react-uswds";
-import * as dateFns from "date-fns";
 import React from "react";
 import parse from "html-react-parser";
 import classNames from "classnames";
+import { Table } from "@trussworks/react-uswds";
+import { AccordionLabResults } from "@/app/view-data/components/AccordionLabResults";
+import {
+  formatAddress,
+  formatDate,
+  formatName,
+  formatPhoneNumber,
+  formatStartEndDateTime,
+  formatVitals,
+  formatDateTime,
+} from "@/app/format-service";
 
 export interface DisplayData {
   title: string;
-  value?: string | React.JSX.Element;
+  value?: string | React.JSX.Element | React.JSX.Element[];
 }
 
 export interface PathMappings {
@@ -27,26 +38,10 @@ export interface ColumnInfoInput {
   infoPath: string;
 }
 
-export const formatPatientName = (
-  fhirBundle: Bundle,
-  fhirPathMappings: PathMappings,
-) => {
-  const givenNames = evaluate(
-    fhirBundle,
-    fhirPathMappings.patientGivenName,
-  ).join(" ");
-  const familyName = evaluate(fhirBundle, fhirPathMappings.patientFamilyName);
-
-  return `${givenNames} ${familyName}`;
-};
-
-const formatName = (firstName: string, lastName: string) => {
-  if (firstName != undefined) {
-    return `${firstName} ${lastName}`;
-  } else {
-    return undefined;
-  }
-};
+export interface CompleteData {
+  availableData: DisplayData[];
+  unavailableData: DisplayData[];
+}
 
 export const extractPatientAddress = (
   fhirBundle: Bundle,
@@ -94,28 +89,6 @@ export const extractFacilityAddress = (
   return formatAddress(streetAddresses, city, state, zipCode, country);
 };
 
-const formatAddress = (
-  streetAddress: string[],
-  city: string,
-  state: string,
-  zipCode: string,
-  country: string,
-) => {
-  let address = {
-    streetAddress: streetAddress || [],
-    cityState: [city, state],
-    zipCodeCountry: [zipCode, country],
-  };
-
-  return [
-    address.streetAddress.join("\n"),
-    address.cityState.filter(Boolean).join(", "),
-    address.zipCodeCountry.filter(Boolean).join(", "),
-  ]
-    .filter(Boolean)
-    .join("\n");
-};
-
 export const extractFacilityContactInfo = (
   fhirBundle: Bundle,
   fhirPathMappings: PathMappings,
@@ -128,206 +101,6 @@ export const extractFacilityContactInfo = (
     (contact: any) => contact.system === "phone",
   );
   return phoneNumbers?.[0].value;
-};
-
-export const formatPatientContactInfo = (
-  fhirBundle: Bundle,
-  fhirPathMappings: PathMappings,
-) => {
-  const phoneNumbers = evaluate(
-    fhirBundle,
-    fhirPathMappings.patientPhoneNumbers,
-  )
-    .map(
-      (phoneNumber) =>
-        `${
-          phoneNumber?.use?.charAt(0).toUpperCase() +
-          phoneNumber?.use?.substring(1)
-        } ${phoneNumber.value}`,
-    )
-    .join("\n");
-  const emails = evaluate(fhirBundle, fhirPathMappings.patientEmails)
-    .map((email) => `${email.value}`)
-    .join("\n");
-
-  return `${phoneNumbers}\n${emails}`;
-};
-
-export const formatEncounterDate = (
-  fhirBundle: Bundle,
-  fhirPathMappings: PathMappings,
-) => {
-  const startDate = formatDateTime(
-    evaluate(fhirBundle, fhirPathMappings.encounterStartDate).join(""),
-  );
-  const endDate = formatDateTime(
-    evaluate(fhirBundle, fhirPathMappings.encounterEndDate).join(""),
-  );
-
-  return `Start: ${startDate}
-    End: ${endDate}`;
-};
-
-const formatDateTime = (dateTime: string) => {
-  const options: Intl.DateTimeFormatOptions = {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "numeric",
-    minute: "2-digit",
-  };
-
-  return new Date(dateTime)
-    .toLocaleDateString("en-Us", options)
-    .replace(",", "");
-};
-
-/**
- * Formats the provided date string into a formatted date string with year, month, and day.
- * @param {string} date - The date string to be formatted.
- * @returns {string | undefined} - The formatted date string or undefined if the input date is falsy.
- */
-export const formatDate = (date?: string): string | undefined => {
-  if (date) {
-    return new Date(date).toLocaleDateString("en-US", {
-      year: "numeric",
-      month: "2-digit",
-      day: "2-digit",
-      timeZone: "UTC",
-    }); // UTC, otherwise will have timezone issues
-  }
-};
-
-const formatPhoneNumber = (phoneNumber: string) => {
-  try {
-    return phoneNumber
-      .replace(/\D/g, "")
-      .replace(/(\d{3})(\d{3})(\d{4})/, "$1-$2-$3");
-  } catch {
-    return undefined;
-  }
-};
-
-const formatStartEndDateTime = (
-  startDateTime: "string",
-  endDateTime: "string",
-) => {
-  const startDateObject = new Date(startDateTime);
-  const endDateObject = new Date(endDateTime);
-
-  const options: Intl.DateTimeFormatOptions = {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "numeric",
-    minute: "numeric",
-    hour12: true,
-  };
-
-  const startFormattedDate = startDateObject
-    .toLocaleString("en-US", options)
-    .replace(",", "");
-  const endFormattedDate = endDateObject
-    .toLocaleString("en-us", options)
-    .replace(",", "");
-
-  return `Start: ${startFormattedDate}
-        End: ${endFormattedDate}`;
-};
-
-const formatVitals = (
-  heightAmount: string,
-  heightMeasurementType: string,
-  weightAmount: string,
-  weightMeasurementType: string,
-  bmi: string,
-) => {
-  let heightString = "";
-  let weightString = "";
-  let bmiString = "";
-
-  let heightType = "";
-  let weightType = "";
-  if (heightAmount && heightMeasurementType) {
-    if (heightMeasurementType === "[in_i]") {
-      heightType = "inches";
-    } else if (heightMeasurementType === "cm") {
-      heightType = "cm";
-    }
-    heightString = `Height: ${heightAmount} ${heightType}\n\n`;
-  }
-
-  if (weightAmount && weightMeasurementType) {
-    if (weightMeasurementType === "[lb_av]") {
-      weightType = "Lbs";
-    } else if (weightMeasurementType === "kg") {
-      weightType = "kg";
-    }
-    weightString = `Weight: ${weightAmount} ${weightType}\n\n`;
-  }
-
-  if (bmi) {
-    bmiString = `Body Mass Index (BMI): ${bmi}`;
-  }
-
-  const combinedString = `${heightString} ${weightString} ${bmiString}`;
-  return combinedString.trim();
-};
-
-/**
- * Formats a table based on the provided resources, mappings, columns, and caption.
- * @param {FhirResource[]} resources - An array of FHIR Resources representing the data entries.
- * @param {PathMappings} mappings - An object containing the FHIR path mappings.
- * @param {ColumnInfoInput[]} columns - An array of objects representing column information.
- *                                      The order of columns in the array determines the order of appearance.
- * @param {string} caption - The caption for the table.
- * @returns {React.JSX.Element} - A formatted table React element.
- */
-const formatTable = (
-  resources: FhirResource[],
-  mappings: PathMappings,
-  columns: ColumnInfoInput[],
-  caption: string,
-): React.JSX.Element => {
-  let headers = columns.map((column, index) => (
-    <th
-      key={`${column.columnName}${index}`}
-      scope="col"
-      className="bg-gray-5 minw-15"
-    >
-      {column.columnName}
-    </th>
-  ));
-
-  let tableRows = resources.map((entry, index) => {
-    let rowCells = columns.map((column, index) => {
-      let rowCellData = evaluate(entry, mappings[column.infoPath])[0] ?? (
-        <span className={"text-italic text-base"}>No data</span>
-      );
-      return (
-        <td key={`row-data-${index}`} className="text-top">
-          {rowCellData}
-        </td>
-      );
-    });
-
-    return <tr key={`table-row-${index}`}>{rowCells}</tr>;
-  });
-
-  return (
-    <Table
-      bordered={false}
-      fullWidth={true}
-      caption={caption}
-      className="border-top border-left border-right table-caption-margin margin-y-0"
-      data-testid="table"
-    >
-      <thead>
-        <tr>{headers}</tr>
-      </thead>
-      <tbody>{tableRows}</tbody>
-    </Table>
-  );
 };
 
 /**
@@ -380,7 +153,7 @@ export const evaluateSocialData = (
   fhirBundle: Bundle,
   mappings: PathMappings,
 ) => {
-  const socialData = [
+  const socialData: DisplayData[] = [
     {
       title: "Occupation",
       value: evaluate(fhirBundle, mappings["patientCurrentJobTitle"])[0],
@@ -421,17 +194,124 @@ export const evaluateSocialData = (
   return evaluateData(socialData);
 };
 
+export const evaluatePatientName = (
+  fhirBundle: Bundle,
+  fhirPathMappings: PathMappings,
+) => {
+  const givenNames = evaluate(
+    fhirBundle,
+    fhirPathMappings.patientGivenName,
+  ).join(" ");
+  const familyName = evaluate(fhirBundle, fhirPathMappings.patientFamilyName);
+
+  return `${givenNames} ${familyName}`;
+};
+
+export const evaluatePatientContactInfo = (
+  fhirBundle: Bundle,
+  fhirPathMappings: PathMappings,
+) => {
+  const phoneNumbers = evaluate(
+    fhirBundle,
+    fhirPathMappings.patientPhoneNumbers,
+  )
+    .map(
+      (phoneNumber) =>
+        `${
+          phoneNumber?.use?.charAt(0).toUpperCase() +
+          phoneNumber?.use?.substring(1)
+        } ${phoneNumber.value}`,
+    )
+    .join("\n");
+  const emails = evaluate(fhirBundle, fhirPathMappings.patientEmails)
+    .map((email) => `${email.value}`)
+    .join("\n");
+
+  return `${phoneNumbers}\n${emails}`;
+};
+export const evaluateEncounterDate = (
+  fhirBundle: Bundle,
+  fhirPathMappings: PathMappings,
+) => {
+  const startDate = formatDateTime(
+    evaluate(fhirBundle, fhirPathMappings.encounterStartDate).join(""),
+  );
+  const endDate = formatDateTime(
+    evaluate(fhirBundle, fhirPathMappings.encounterEndDate).join(""),
+  );
+
+  return `Start: ${startDate}
+    End: ${endDate}`;
+};
+/**
+ * Formats a table based on the provided resources, mappings, columns, and caption.
+ * @param {FhirResource[]} resources - An array of FHIR Resources representing the data entries.
+ * @param {PathMappings} mappings - An object containing the FHIR path mappings.
+ * @param {ColumnInfoInput[]} columns - An array of objects representing column information.
+ *                                      The order of columns in the array determines the order of appearance.
+ * @param {string} caption - The caption for the table.
+ * @returns {React.JSX.Element} - A formatted table React element.
+ */
+export const evaluateTable = (
+  resources: FhirResource[],
+  mappings: PathMappings,
+  columns: ColumnInfoInput[],
+  caption: string,
+): React.JSX.Element => {
+  let headers = columns.map((column, index) => (
+    <th
+      key={`${column.columnName}${index}`}
+      scope="col"
+      className="bg-gray-5 minw-15"
+    >
+      {column.columnName}
+    </th>
+  ));
+
+  let tableRows = resources.map((entry, index) => {
+    let rowCells = columns.map((column, index) => {
+      let rowCellData = evaluate(entry, mappings[column.infoPath])[0] ?? (
+        <span className={"text-italic text-base"}>No data</span>
+      );
+      return (
+        <td key={`row-data-${index}`} className="text-top">
+          {rowCellData}
+        </td>
+      );
+    });
+
+    return <tr key={`table-row-${index}`}>{rowCells}</tr>;
+  });
+
+  return (
+    <Table
+      bordered={false}
+      fullWidth={true}
+      caption={caption}
+      className="border-top border-left border-right table-caption-margin margin-y-0"
+      data-testid="table"
+    >
+      <thead>
+        <tr>{headers}</tr>
+      </thead>
+      <tbody>{tableRows}</tbody>
+    </Table>
+  );
+};
 export const evaluateDemographicsData = (
   fhirBundle: Bundle,
   mappings: PathMappings,
 ) => {
-  const demographicsData = [
+  const demographicsData: DisplayData[] = [
     {
       title: "Patient Name",
-      value: formatPatientName(fhirBundle, mappings),
+      value: evaluatePatientName(fhirBundle, mappings),
     },
     { title: "DOB", value: evaluate(fhirBundle, mappings.patientDOB)[0] },
-    { title: "Current Age", value: calculatePatientAge(fhirBundle, mappings) },
+    {
+      title: "Current Age",
+      value: calculatePatientAge(fhirBundle, mappings).toString(),
+    },
     { title: "Sex", value: evaluate(fhirBundle, mappings.patientGender)[0] },
     { title: "Race", value: evaluate(fhirBundle, mappings.patientRace)[0] },
     {
@@ -454,7 +334,10 @@ export const evaluateDemographicsData = (
       title: "County",
       value: evaluate(fhirBundle, mappings.patientCounty)[0],
     },
-    { title: "Contact", value: formatPatientContactInfo(fhirBundle, mappings) },
+    {
+      title: "Contact",
+      value: evaluatePatientContactInfo(fhirBundle, mappings),
+    },
     {
       title: "Emergency Contact",
       value: evaluateEmergencyContact(fhirBundle, mappings),
@@ -581,7 +464,7 @@ export const evaluateEcrMetadata = (
     },
     {
       title: "Sender Facility Name",
-      value: evaluate(fhirBundle, mappings.senderFacilityName),
+      value: evaluate(fhirBundle, mappings.senderFacilityName)[0],
     },
     {
       title: "Facility Address",
@@ -633,7 +516,7 @@ export const returnProblemsTable = (
       new Date(a.onsetDateTime ?? "").getTime(),
   );
 
-  return formatTable(problemsArray, mappings, columnInfo, "Problems List");
+  return evaluateTable(problemsArray, mappings, columnInfo, "Problems List");
 };
 
 /**
@@ -666,7 +549,7 @@ export const returnImmunizations = (
       new Date(a.occurrenceDateTime ?? "").getTime(),
   );
 
-  return formatTable(
+  return evaluateTable(
     immunizationsArray,
     mappings,
     columnInfo,
@@ -704,7 +587,7 @@ export const returnProceduresTable = (
       new Date(a.performedDateTime ?? "").getTime(),
   );
 
-  return formatTable(proceduresArray, mappings, columnInfo, "Procedures");
+  return evaluateTable(proceduresArray, mappings, columnInfo, "Procedures");
 };
 
 export const evaluateClinicalData = (
@@ -799,19 +682,6 @@ const evaluateData = (
   return { availableData: availableData, unavailableData: unavailableData };
 };
 
-export const formatString = (input: string): string => {
-  // Convert to lowercase
-  let result = input.toLowerCase();
-
-  // Replace spaces with underscores
-  result = result.replace(/\s+/g, "-");
-
-  // Remove all special characters except underscores
-  result = result.replace(/[^a-z0-9\-]/g, "");
-
-  return result;
-};
-
 /**
  * Functional component for displaying data.
  * @param {object} props - Props for the component.
@@ -897,4 +767,53 @@ export const evaluateEmergencyContact = (
 
     return formattedContact;
   }
+};
+
+/**
+ * Evaluates lab information and RR data from the provided FHIR bundle and mappings.
+ * @param {Bundle} fhirBundle - The FHIR bundle containing lab and RR data.
+ * @param {PathMappings} mappings - An object containing the FHIR path mappings.
+ * @returns {{
+ *   labInfo: CompleteData,
+ *   labResults: React.JSX.Element[]
+ * }} An object containing evaluated lab information and lab results.
+ */
+export const evaluateLabInfoData = (
+  fhirBundle: Bundle,
+  mappings: PathMappings,
+): {
+  labInfo: CompleteData;
+  labResults: React.JSX.Element[];
+} => {
+  const labInfo: DisplayData[] = [
+    {
+      title: "Lab Performing Name",
+      value: "",
+    },
+    {
+      title: "Lab Address",
+      value: "",
+    },
+    {
+      title: "Lab Contact",
+      value: "",
+    },
+  ];
+
+  const rrData = evaluate(fhirBundle, mappings["diagnosticReports"]).map(
+    (report) => {
+      return (
+        <AccordionLabResults
+          title={report.code.coding[0].display}
+          abnormalTag={false}
+          content={<></>}
+        />
+      );
+    },
+  );
+
+  return {
+    labInfo: evaluateData(labInfo),
+    labResults: rrData,
+  };
 };
